// source/frontend/src/models/api/generated/ApiGen_Concepts_LeasePurpose.ts
/**
 * File autogenerated by TsGenerator.
 * Do not manually modify, changes made to this file will be lost when this file is regenerated.
 */
import { ApiGen_Base_BaseAudit } from './ApiGen_Base_BaseAudit';
import { ApiGen_Base_CodeType } from './ApiGen_Base_CodeType';

// LINK: @backend/apimodels/Models/Concepts/Lease/LeasePurposeModel.cs
export interface ApiGen_Concepts_LeasePurpose extends ApiGen_Base_BaseAudit {
  id: number;
  leaseId: number;
  leasePurposeTypeCode: ApiGen_Base_CodeType<string> | null;
  purposeOtherDescription: string | null;
}